refactor(dashboard): extract basic shape
