+++ src/containers/thread/DashboardThread/UI/PostListLayout.tsx
@@ -5,7 +5,7 @@ import type { TPostLayout } from '@/spec'
 import { POST_LAYOUT, DASHBOARD_DESC_LAYOUT } from '@/constant'
 import { callDashboardDesc } from '@/utils/helper'
 
-import { Br, Space, SpaceGrow, Inline } from '@/widgets/Common'
+import { Row, Br, Space, SpaceGrow, Inline } from '@/widgets/Common'
 import ArrowButton from '@/widgets/Buttons/ArrowButton'
 import CheckLabel from '@/widgets/CheckLabel'
 
@@ -21,7 +21,6 @@ import {
   Block,
   Bar,
   Circle,
-  Row,
   Column,
   UpvoteIcon,
   CommentIcon,
