+++ src/containers/thread/DashboardThread/styles/ui/index.ts
@@ -37,3 +37,20 @@ export const BlockBase = styled.div<TActive>`
 
   transition: all 0.2s;
 `
+
+// base shapes
+type TBar = { long: number; thin?: boolean }
+export const Bar = styled.div<TBar>`
+  width: ${({ long }) => `${long || 10}%`};
+  height: ${({ thin }) => (thin ? '4px' : '10px;')};
+  background: ${({ thin }) =>
+    thin ? theme('thread.articleDigest') : theme('thread.articleTitle')};
+  z-index: 3;
+  border-radius: 5px;
+  opacity: ${({ thin }) => (thin ? 0.6 : 1)};
+`
+
+export const Circle = styled.div<{ radius?: number }>`
+  ${({ radius }) => `${css.circle(radius || 22)}`};
+  background: ${theme('thread.articleTitle')};
+`
